% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSTCK.R
\name{SRSTCK}
\alias{SRSTCK}
\title{SRSTCK main_func.}
\usage{
SRSTCK <- function (data, cancer, type, cell_type, multi.sample = "FALSE", minGene = 500,
                    maxGene = 5000, maxUMI = 40000, pctMT = 20, id.type = "S",
                    cell.line = "no", ngene.chr = 5, win.size = 25, KS.cut = 0.15,
                    distance = "euclidean", n.cores = 1)

SRSTCK(
  data,
  cancer,
  type,
  cell_type,
  multi.sample = "FALSE",
  minGene = 500,
  maxGene = 5000,
  maxUMI = 40000,
  pctMT = 20,
  id.type = "S",
  cell.line = "no",
  ngene.chr = 5,
  win.size = 25,
  KS.cut = 0.15,
  distance = "euclidean",
  n.cores = 1
)
}
\arguments{
\item{data}{raw data matrix; if the data is scRNA-seq, genes in rows; cell names in columns; if the data is bulk, genes in rows; sample names in columns.}

\item{cancer}{short for cancer: such as "OVCA";"CRC";SKCM" etc.}

\item{type}{if the data is scRNA-seq, put "sc"; if the data is bulk RNA-seq, put "bulk".}

\item{cell_type}{if the data is scRNA-seq, "cell_type" represents a collection of TME cells, including immune cells, fibroblasts, and endothelial cells. Please convert these cell types in your data into a vector and pass it to "cell_type".}

\item{multi.sample}{if scRNA-seq is a single sample, put "FALSE"; if scRNA-seq is multi-sample integration, put "TRUE".}

\item{minGene}{the minimum number of genes in a cell.}

\item{maxGene}{the maximum number of genes in a cell.}

\item{maxUMI}{the maximum UMI count in a cell.}

\item{pctMT}{mitochondrial gene ratio.}

\item{id.type}{gene id type: Symbol or Ensemble.}

\item{cell.line}{if the data are from pure cell line,put "yes"; if cell line data are a mixture of tumor and normal cells, still put "no".}

\item{ngene.chr}{minimal number of genes per chromosome for cell filtering.}

\item{win.size}{minimal window sizes for segmentation.}

\item{KS.cut}{segmentation parameters, input 0 to 1; larger looser criteria.}

\item{distance}{distance methods include euclidean, and correlation converted distance include pearson and spearman.}

\item{n.cores}{number of cores for parallel computing.}
}
\value{
if data is scRNA-seq: 1) aneuploid/diploid prediction results; 2)senescent tumor cells prediction results; 3)annotated cluster diagram of senescent tumor cells; if data is bulk: the senescence score results.
}
\description{
SRSTCK main_func.
}
\examples{
test.SRSTCK <-  SRSTCK(data = data, cancer = "CRC", type = "sc", cell_type = TME_cell_types, multi.sample = "FALSE", minGene = 500, maxGene = 5000, maxUMI = 40000, pctMT = 20, id.type = "S", cell.line = "no", ngene.chr = 5,  win.size = 25, KS.cut = 0.15, distance = "euclidean", n.cores = 1)
}
